pyproject.toml
[project]
name = "chatbot-new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "openai>=1.3.0",
    "pydantic==2.5.0",
    "python-dotenv>=1.0.0",
    "python-multipart==0.0.6",
    "requests>=2.31.0",
    "uvicorn[standard]==0.24.0",
]
